+++ packages/editor/src/deserializeHtml/tests/html-samples/elements.ts
@@ -1,0 +1,40 @@
+export default `<p>The <b>brown-eared bulbul</b> (<i>Hypsipetes amaurotis</i>) is a medium-sized <a href="/wiki/Bulbul" title="Bulbul">bulbul</a> native to eastern Asia. It is extremely common within the northern parts of its range and can be found from southern <a href="/wiki/Sakhalin" title="Sakhalin">Sakhalin</a> to the northern Philippines.
+</p>
+<div id="toc" class="toc" role="navigation" aria-labelledby="mw-toc-heading"><input type="checkbox" role="button" id="toctogglecheckbox" class="toctogglecheckbox" style="display:none"><div class="toctitle" lang="en" dir="ltr"><h2 id="mw-toc-heading">Contents</h2><span class="toctogglespan"><label class="toctogglelabel" for="toctogglecheckbox"></label></span></div>
+<ul>
+<li class="toclevel-1 tocsection-1"><a href="#Taxonomy_and_systematics"><span class="tocnumber">1</span> <span class="toctext">Taxonomy and systematics</span></a>
+<ul>
+<li class="toclevel-2 tocsection-2"><a href="#Subspecies"><span class="tocnumber">1.1</span> <span class="toctext">Subspecies</span></a></li>
+</ul>
+</li>
+<li class="toclevel-1 tocsection-3"><a href="#Description"><span class="tocnumber">2</span> <span class="toctext">Description</span></a></li>
+<li class="toclevel-1 tocsection-4"><a href="#Distribution_and_habitat"><span class="tocnumber">3</span> <span class="toctext">Distribution and habitat</span></a></li>
+<li class="toclevel-1 tocsection-5"><a href="#Behaviour"><span class="tocnumber">4</span> <span class="toctext">Behaviour</span></a>
+<ul>
+<li class="toclevel-2 tocsection-6"><a href="#Breeding"><span class="tocnumber">4.1</span> <span class="toctext">Breeding</span></a></li>
+<li class="toclevel-2 tocsection-7"><a href="#Feeding"><span class="tocnumber">4.2</span> <span class="toctext">Feeding</span></a></li>
+</ul>
+</li>
+<li class="toclevel-1 tocsection-8"><a href="#References"><span class="tocnumber">5</span> <span class="toctext">References</span></a></li>
+<li class="toclevel-1 tocsection-9"><a href="#Further_reading"><span class="tocnumber">6</span> <span class="toctext">Further reading</span></a></li>
+<li class="toclevel-1 tocsection-10"><a href="#External_links"><span class="tocnumber">7</span> <span class="toctext">External links</span></a></li>
+</ul>
+</div>
+<h2><span class="mw-headline" id="Taxonomy_and_systematics">Taxonomy and systematics</span><span class="mw-editsection"><span class="mw-editsection-bracket">[</span><a href="/w/index.php?title=Brown-eared_bulbul&amp;action=edit&amp;section=1" title="Edit section: Taxonomy and systematics">edit</a><span class="mw-editsection-bracket">]</span></span></h2>
+<p>The brown-eared bulbul was originally described in the genus <i><a href="/wiki/Turdus" class="mw-redirect" title="Turdus">Turdus</a>.</i> Later, some authorities placed it in the genus <i><a href="/wiki/Ixos" title="Ixos">Ixos</a></i>  and then the genus <i><a href="/wiki/Microscelis" title="Microscelis">Microscelis</a>,</i> before being re-classified to <i>Hypsipetes</i> in 2010.<sup id="cite_ref-3" class="reference"><a href="#cite_note-3">[3]</a></sup> Alternate names for the brown-eared bulbul include the <b>Asian brown-eared bulbul</b>, <b>chestnut-eared bulbul</b>, and <b>Eurasian brown-eared bulbul</b>.
+</p>
+<h3><span class="mw-headline" id="Subspecies">Subspecies</span><span class="mw-editsection"><span class="mw-editsection-bracket">[</span><a href="/w/index.php?title=Brown-eared_bulbul&amp;action=edit&amp;section=2" title="Edit section: Subspecies">edit</a><span class="mw-editsection-bracket">]</span></span></h3>
+<p>Twelve subspecies are currently recognized:
+</p>
+<ul><li><b>Japanese brown-eared bulbul</b> (<i>H. a. amaurotis)</i> – <small>(<a href="/wiki/Coenraad_Jacob_Temminck" title="Coenraad Jacob Temminck">Temminck</a>, 1830)</small>: Also named the <b>Japanese chestnut-eared bulbul</b>. Found on southern Sakhalin, Japan and South Korea</li>
+<li><i>H. a. matchiae</i> – <small>(<a href="/w/index.php?title=Tokutaro_Momiyama&amp;action=edit&amp;redlink=1" class="new" title="Tokutaro Momiyama (page does not exist)">Momiyama</a>, 1923)</small>: Found on southern <a href="/wiki/Kyushu" title="Kyushu">Kyushu</a> (Japan)</li>
+<li><i>H. a. ogawae</i> – <small><a href="/wiki/Ernst_Hartert" title="Ernst Hartert">Hartert</a>, 1907</small>: Found on northern <a href="/wiki/Ryukyu_Islands" title="Ryukyu Islands">Ryukyu Islands</a></li>
+<li><i>H. a. pryeri</i> – <small><a href="/wiki/Leonhard_Stejneger" title="Leonhard Stejneger">Stejneger</a>, 1887</small>: Found on central Ryukyu Islands</li>
+<li><i>H. a. stejnegeri</i> – <small>Hartert, 1907</small>: Found on southern Ryukyu Islands</li>
+<li><i>H. a. squamiceps</i> – <small>(<a href="/wiki/Heinrich_von_Kittlitz" title="Heinrich von Kittlitz">Kittlitz</a>, 1830)</small>: Originally described as a separate species in the genus <i><a href="/wiki/Oriolus" title="Oriolus">Oriolus</a>.</i> Found on the <a href="/wiki/Bonin_Islands" title="Bonin Islands">Bonin Islands</a></li>
+<li><i>H. a. magnirostris</i> – <small>Hartert, 1905</small>: Found on the <a href="/wiki/Volcano_Islands" title="Volcano Islands">Volcano Islands</a></li>
+<li><i>H. a. borodinonis</i> – <small>(<a href="/wiki/Nagamichi_Kuroda" title="Nagamichi Kuroda">Kuroda</a>, 1923)</small>: Found on the <a href="/wiki/Borodino_Islands" class="mw-redirect" title="Borodino Islands">Borodino Islands</a></li>
+<li><b>Taiwan brown-eared bulbul</b> (<i>H. a. nagamichii)</i> – <small><a href="/wiki/Austin_L._Rand" title="Austin L. Rand">Rand</a> &amp; <a href="/wiki/Herbert_Girton_Deignan" title="Herbert Girton Deignan">Deignan</a>, 1960</small>: Also named the <b>Taiwan chestnut-eared bulbul</b>. Found on Taiwan and <a href="/wiki/Orchid_Island" title="Orchid Island">Lanyu Island</a></li>
+<li><i>H. a. batanensis</i> – <small><a href="/wiki/Edgar_Alexander_Mearns" title="Edgar Alexander Mearns">Mearns</a>, 1907</small>: Found on <a href="/wiki/Batan_Island" title="Batan Island">Batan</a>, <a href="/wiki/Ivuhos" title="Ivuhos">Ivuhos</a>, <a href="/wiki/Sabtang,_Batanes" title="Sabtang, Batanes">Sabtang</a>, <a href="/wiki/Babuyan_Island" title="Babuyan Island">Babuyan</a> and Claro (extreme northern Philippines)</li>
+<li><i>H. a. fugensis</i> – <small><a href="/wiki/William_Robert_Ogilvie-Grant" title="William Robert Ogilvie-Grant">Ogilvie-Grant</a>, 1895</small>: Found on <a href="/wiki/Dalupiri_Island_(Cagayan)" title="Dalupiri Island (Cagayan)">Dalupiri</a>, <a href="/wiki/Calayan,_Cagayan" title="Calayan, Cagayan">Calayan</a> and <a href="/wiki/Fuga_Island" title="Fuga Island">Fuga Island</a> (north of <a href="/wiki/Luzon" title="Luzon">Luzon</a> in northern Philippines)</li>
+<li><i>H. a. camiguinensis</i> – <small><a href="/w/index.php?title=Richard_Crittenden_McGregor&amp;action=edit&amp;redlink=1" class="new" title="Richard Crittenden McGregor (page does not exist)">McGregor</a>, 1907</small>: Found on <a href="/wiki/Camiguin" title="Camiguin">Camiguin Island</a>  (north of Luzon in northern Philippines)</li></ul>`;