+++ packages/editor/src/deserializeHtml/utils/deserializeBreak.ts
@@ -1,0 +1,2 @@
+export default (node: HTMLElement | ChildNode): string | null =>
+  node.nodeName === 'BR' ? '\n' : null;