+++ packages/editor/src/index.ts
@@ -3,3 +3,5 @@ export * from './Editor';
 
 export { default as withPlugins } from './withPlugins';
 export * from './withPlugins';
+
+export { default as Transforms } from './transforms';