+++ packages/editor/src/transforms/index.ts
@@ -1,0 +1,5 @@
+import { Transforms as SlateTransforms } from 'slate';
+
+export default {
+  ...SlateTransforms,
+};