+++ packages/editor/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This is a list of changes to Slash with each new release. Until 1.0.0 is released, breaking changes will be added as minor or patch version bumps.
 
+### `0.3.0` — July 20, 2020
+
+###### NEW
+
+- Add `elementDeserializers` interface to `SlashPlugin` for quickly defining a plugin's element deserializers.
+- Add `markDeserializers` interface to `SlashPlugin` for quickly defining a plugin's mark deserializers.
+
 ### `0.2.0` — May 28, 2020
 
 ###### NEW